+++ src/index.html
@@ -1,0 +1,64 @@
+<!doctype html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.3.0/font/bootstrap-icons.css" />
+    <link href="../dist/output.css" rel="stylesheet">
+    <link
+        href="https://fonts.googleapis.com/css2?family=Caveat&family=Inter:wght@400;500;700&family=Lato:wght@300&family=Ms+Madi&family=Poppins:wght@200&display=swap"
+        rel="stylesheet">
+</head>
+
+<body class="font-[roboto] bg-[url('./img/bg-g.svg')] bg-no-repeat bg-cover">
+    <main class="bg-white mx-auto border-2  max-w-[1198px] my-[2rem] p-[70px] overflow-y-hidden">
+        <div class="text-left text-slate-400">
+            <p class="mb-6 text-left text-3xl font-extrabold p-1">manifest</p>
+        </div>
+        <div class="flex flex-col items-center justify-center">
+            <div class="max-w-md w-full  bg-gray-50  p-8 space-y-4 shadow-md">
+                <div class="text-left text-slate-800">
+                    <p class=" text-left text-md p-1">Nice to see you again!👋</p>
+                </div>
+                <div class="flex flex-col">
+                    <label class="mb-1 text-sm" for="login">Login</label>
+                    <input class="border rounded-md px-3 py-2" type="text" name="login" id="login-in"
+                        Placeholder="Email or Phone Number">
+                </div>
+                <div>
+                    <div class="flex flex-col">
+                        <label class="mb-1 text-sm" for="password">Password</label>
+                        <input class="border rounded-md px-3 py-2 relative" type="password" name="login" id="pass-in"
+                            Placeholder="Enter Password">
+                        <i class="bi bi-eye-slash absolute inset-y-0 right-0" id="togglePassword"></i>
+                    </div>
+                </div>
+                <div class="flex justify-between text-xs">
+                    <div class="flex items-center space-x-2">
+                        <input type="checkbox" name="remember_me" id="remember_me">
+                        <label for="remember_me">Remember Me</label>
+                    </div>
+                    <div>
+                        <a href="#">forgot password?</a>
+                    </div>
+                </div>
+
+                <div>
+                    <button class="w-full bg-[#098B6C] rounded text-white py-2 cursor-pointer">Sign-in</button>
+                </div>
+                <div>
+                    <button class="bg-[#F2F2F2] w-full rounded text-gray-800 py-2  cursor-pointer"> G or Sign-in with
+                        Google</button>
+                </div>
+                <div class="text-center">
+                    <p class=" px-3 text-gray-500 text-xs">Do you have an account?
+                        <button class="text-[#098B6C] cursor-pointer ">Sign-up now</button>
+                    </p>
+                </div>
+            </div>
+        </div>
+    </main>
+</body>
+
+</html>
